Facelift

+++ r2048/Library/UIColor+RTTFromHex.m
@@ -12,9 +12,9 @@
 }
 
 + (UIColor *)fromHex:(NSUInteger)rgbValue alpha:(float)alpha {
-    return [UIColor colorWithRed:((float)((rgbValue & 0xFF0000) >> 16)) / 255.0
-                           green:((float)((rgbValue & 0xFF00) >> 8)) / 255.0
-                            blue:((float)(rgbValue & 0xFF)) / 255.0
+    return [UIColor colorWithRed:((rgbValue & 0xFF0000) >> 16) / 255.0f
+                           green:((rgbValue & 0xFF00) >> 8) / 255.0f
+                            blue:(rgbValue & 0xFF) / 255.0f
                            alpha:alpha];
 }
